+++ vrp-cli/src/extensions/solve/config.rs
@@ -331,7 +331,8 @@ pub struct TerminationConfig {
 #[derive(Clone, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct VariationConfig {
-    sample: usize,
+    interval_type: String,
+    value: usize,
     cv: f64,
     is_global: bool,
 }
@@ -519,7 +520,8 @@ fn configure_from_termination(mut builder: Builder, termination_config: &Option<
     if let Some(config) = termination_config {
         builder = builder.with_max_time(config.max_time);
         builder = builder.with_max_generations(config.max_generations);
-        builder = builder.with_min_cv(config.variation.as_ref().map(|v| (v.sample, v.cv, v.is_global)));
+        builder = builder
+            .with_min_cv(config.variation.as_ref().map(|v| (v.interval_type.clone(), v.value, v.cv, v.is_global)));
     }
 
     builder